Marshal support for PipelineDB's HyperLogLog structure.
(As compiled on x64 and observed on database-2.)

Mimic the HyperLogLog structure and serialize a retailnext HLLPP object
to this structure. It needs to have the same padding as the one observed
on database-2. Ideally, the PipelineDB version should be packed and
versioned, but doing so would invalidate our existing HyperLogLogs in
the database. For now, we'll just match the one that we use.

Issue arbortech/workspace#1437