Fix sparse mode flush issue
When sorting the temp set before merging with the (already sorted)
sparse registers, we were sorting by the plain uint32 value, but that
is not correct in all cases. The intent is to sort by the "index"
encoded in the uint32 value, but the bit offset of the index is not
consistent (the rho value is not encoded in the uint32 if not needed,
which shifts the msb of the index by 6, depending on whether rho is
there or not). Now, we explicitly decode and sort by the index inside
the uint32.

The impact of this bug was elevated estimate error for very low
cardinalities.